add admin controllers for product's listing, deletion, creation and updation

+++ backend/src/lib/imagekit.ts
@@ -1,15 +1,17 @@
 import ImageKit from "@imagekit/nodejs";
 import { NotFoundError } from "@imagekit/nodejs";
 
-export async function deleteImageFromImagekit(imagekitFileId: string){
+export async function deleteImageFromImagekit(imagekitFileId: string) {
     try {
+        if (!imagekitFileId) return
+
         const client = new ImageKit({
             privateKey: process.env.IMAGEKIT_PRIVATE_KEY,
         })
 
         await client.files.delete(imagekitFileId);
     } catch (error) {
-        if(error instanceof NotFoundError) return
+        if (error instanceof NotFoundError) return
         throw error
     }
 }
